+++ ranger-http-server/src/main/java/io/appform/ranger/http/server/HttpAppConfiguration.java
@@ -15,6 +15,7 @@
  */
 package io.appform.ranger.http.server;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.appform.ranger.http.server.bundle.config.RangerHttpConfiguration;
 import io.dropwizard.Configuration;
 import lombok.AllArgsConstructor;
@@ -30,6 +31,7 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class HttpAppConfiguration extends Configuration {
     @NotEmpty
     @NotNull