+++ ranger-zookeeper/src/test/java/io/appform/ranger/zookeeper/model/ServiceProviderTest.java
@@ -108,7 +108,7 @@ class ServiceProviderTest {
     }
 
     @Test
-    public void testBasicDiscoveryRR() {
+    void testBasicDiscoveryRR() {
         val serviceFinder
                 = ServiceFinderBuilders.<TestNodeData>shardedFinderBuilder()
                 .withConnectionString(testingCluster.getConnectString())